moved ExpressionParser from AbstractTestClass to its own class to remove dependencies to test packages.
