+++ plugins/org.yakindu.sct.model.stext/META-INF/MANIFEST.MF
@@ -43,6 +43,7 @@ Export-Package: org.yakindu.sct.model.stext,
  org.yakindu.sct.model.stext.stext,
  org.yakindu.sct.model.stext.stext.impl,
  org.yakindu.sct.model.stext.stext.util,
+ org.yakindu.sct.model.stext.expressions,
  org.yakindu.sct.model.stext.util,
  org.yakindu.sct.model.stext.validation
 Bundle-ClassPath: .
